requirements.txt
# Kafka and Streaming
confluent-kafka==2.3.0
kafka-python==2.0.2

# Spark and Big Data
pyspark==3.5.0
pyarrow==14.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2

# Data Generation and Validation
faker==20.1.0
jsonschema==4.20.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Airflow (for DAG execution)
apache-airflow==2.8.0
